app_ai/requirements.txt
# FastAPI 및 서버 실행
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6

# 데이터베이스
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
python-dotenv>=1.0.1

# AI/ML 라이브러리 (무거운 의존성)
torch>=2.0.0
transformers>=4.30.0
librosa>=0.10.0
scipy>=1.10.0

# Google Cloud 서비스
google-cloud-speech>=2.21.0
google-cloud-language>=2.8.0
google-auth>=2.23.0

# OpenAI
openai>=1.40.0
